--- src/boilerdata/models.py
@@ -1,11 +1,0 @@
-from enum import auto
-
-from boilerdata.enums import NameEnum
-from boilerdata.trials import Trials
-
-
-class Model(NameEnum):
-    Trials = auto()
-
-
-model_from_cli = {Model.Trials: Trials}
