+++ src/boilerdata/models/__init__.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class EnumValueBaseModel(BaseModel):
+    class Config:
+        use_enum_values = True
